Move language selector to components
