+++ django_project/frontend/src/components/LanguageSelector/style.scss
